remove version - replace by buildinfo
