--- lib/version/version.cpp
@@ -1,21 +1,0 @@
-#include <version.hpp>
-#include <buildinfo.hpp>
-#include <stdio.h>
-
-uint8_t version::isVersionMajor{0};
-uint8_t version::isVersionMinor{0};
-uint8_t version::isVersionPatch{0};
-buildType version::theBuildType{buildType::unknown};
-buildEnvironment version::theBuildEnvironment{buildEnvironment::unknown};
-
-char version::isVersionString[maxVersionStringLength]{};
-
-void version::initialize() {
-    isVersionMajor      = buildInfo::mainVersionDigit;
-    isVersionMinor      = buildInfo::minorVersionDigit;
-    isVersionPatch      = buildInfo::patchVersionDigit;
-    theBuildType        = buildInfo::theBuildType;
-    theBuildEnvironment = buildInfo::theBuildEnvironment;
-    snprintf(isVersionString, maxVersionStringLength, "v%u.%u.%u", isVersionMajor, isVersionMinor, isVersionPatch);
-}
-
